Security Filter Chain updated for corsconfig
