+++ Backend/src/main/java/student_info/configuration/WebConfig.java
@@ -1,17 +1,17 @@
-package student_info.configuration;
-
-//import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**") 
-                .allowedOrigins("*") 
-                .allowedMethods("*") 
-                .allowedHeaders("*"); 
-    }
-}
+//package student_info.configuration;
+//
+////import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.web.servlet.config.annotation.CorsRegistry;
+//import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+//
+//@Configuration
+//public class WebConfig implements WebMvcConfigurer {
+//    @Override
+//    public void addCorsMappings(CorsRegistry registry) {
+//        registry.addMapping("/**") 
+//                .allowedOrigins("*") 
+//                .allowedMethods("*") 
+//                .allowedHeaders("*"); 
+//    }
+//}
